Allow watering can to be used on beets
This is very much a learning commit for me. I'm trying to add the ability for a watering can to be used on Dry Planted Beets. I'm just making an educated guess on what to edit though. Still learning how objects, transitions, and categories work together.

Also, my process was just to search through twotech and the codebase looking for object ID's. If there's a better way to do this I'd love to know.
I'd also love to know how to test. Is there a dev server that's running somewhere? Is it tricky to do a localhost server instance?