Consistent orientation and containability of bowls of stuff when contained
generally, if it is not liquid, it is containable as large item, and it should not be flipped sideway when stored in containers like wooden box. Else if it is a bowl of food, not containable except table/shelf; else if there are some other ways to store it like buckets/bottles etc it should not be containable; else it could be containable.